requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.29.0
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wsproto==1.2.0
